trying to make plotting robust for Mixed models
For the mixed models we have group-level parameters (i.e. the
unobserved participant), but only for SOME parameters. Turns out I’ve
implicitly assumed we either have no or all parameters for at the
group-level, and have coded this assumption in. So I’m attempting to
make things robust to this situation.

Also tidied up sampler wrappers a bit